Renamed module dummy_bus.rs to dummy.rs
